This is how I'm going to be writing .platform files from now on.
The @commonplatforms is crufty and confusing -- it is functionally equivalent
to adding the -I%T/platforms/blah yourself.  But, I always thought it did
more, like bring it the other .platform files, as well.  But it doesn't.

I'm also using the @opts=qw() format because it saves on quotes and commas,
so maybe it's more clear?

Have at thee!