TOS standard 32-bit interfaces are now just Alarm and Counter.
Interfaces templated with size are now AlarmBase and CounterBase.

MathOps interface created to allow for mathmatical ops on template types.

Alarm and AlarmBase interfaces now include command now() to get the current
time.  This is a necessary idiom for Alarms with the anchored-relative (t0,dt)
alarm specifications, so it has been included in the alarm interface instead of
requiring a counter be wired in, as well.

clearOverflow is now void instead of returning a bool.