TOS_LOCAL_ADDRESS -> TOS_NODE_ID
Now, when you make install.x, you set both the AM address and the NODE_ID
of a mote. The former is used for AM communication, the latter is used for
random seeds. Eventually the latter should be replaced with chip IDs and put
behind an interface, but for now it is fine.