Fixed two bugs in packet simulation.
The first was documented by Avinash from USC. He noticed that in some
cases of packet corruption a node would never return to searching
for new packets (it was stuck in a data-receive state). This would
cause it to stop receiving packets. The fix was setting receiving to 0
in an additional case.

The second had to do with properly simulating the radio channel when
some nodes were not yet on. This required letting nodes handle events
that simulate its channel when off without receiving packets. Fixing
this also tickled a few other minor bugs (e.g., field initialization).