Utility for bounded dynamic memory allocation.
FragmentPoolC manages a large buffer from which components may request
blocks of data of varying size.  It differs from PoolC in that it is not
tied to a particular data type, and because it's intended to hold incoming
messages of unknown size but return the unused portion of an allocated block
to the pool while the used portion is retained.  In essence: malloc.

This branch includes a detailed unit test, and so depends on osian/unittest.